Rewrite with private instance data and auto shim
Rewrite to use the best available implementation through detection of
properties on prototypes of Document and NodeIterator.

Be paranoid about function binding, as directed by the es-shim API
documentation.

Make attributes of the wrapper and the sham implementation readonly and
keep private data out of reach.

Make wrapper objects instances of NodeIterator by setting up their
prototype chains.

Add a "dom-node-iterator/auto" entrypoint that automatically installs a
shim when the builtin implementation is not spec-compliant.